+++ src/main/java/com/mraof/minestuck/data/MinestuckItemModelProvider.java
@@ -24,6 +24,16 @@ public class MinestuckItemModelProvider extends ItemModelProvider
 	{
 		
 		handheldItemTextureName(MSItems.ACE_OF_CLUBS, "ace_clubs");
+
+		//Staffs
+		handheldItemTextureName(MSItems.WIZARD_STAFF, "wizard_staff");
+		handheldItemTextureName(MSItems.WATER_STAFF, "water_staff");
+		handheldItemTextureName(MSItems.FIRE_STAFF, "fire_staff");
+
+		//Clubs
+		handheldItem(MSItems.M_ACE_OF_CLUBS);
+		handheldItem(MSItems.HOME_GROWN_MACE);
+		handheldItemTextureName(MSItems.RUBIKS_MACE, "rubiks_mace");
 		
 		//Blocks
 		blockItem(MSBlocks.UNCARVED_WOOD);